Moved layout code
